style : Error타입들을 폴더하나로 묶음처리
