+++ Calculator/Calculator/Errors/CalculatorItemQueueError.swift
@@ -1,0 +1,12 @@
+//
+//  CalculatorItemQueueError.swift
+//  Calculator
+//
+//  Created by 이원빈 on 2022/05/25.
+//
+
+import Foundation
+
+enum CalculatorItemQueueError: Error {
+    case emptyStack
+}
