+++ Calculator/Calculator/Errors/FormulaError.swift
@@ -1,0 +1,12 @@
+//
+//  FormulaError.swift
+//  Calculator
+//
+//  Created by 이원빈 on 2022/05/25.
+//
+
+import Foundation
+
+enum FormulaError: Error {
+    case notEnoughInput
+}
