+++ Calculator/Calculator/Errors/OperatorError.swift
@@ -1,0 +1,12 @@
+//
+//  OperatorError.swift
+//  Calculator
+//
+//  Created by 이원빈 on 2022/05/25.
+//
+
+import Foundation
+
+enum OperatorError: Error {
+    case dividedByZero
+}
